backend/requirements.txt
fastapi
uvicorn
python-dotenv
twilio
python-multipart
pydantic